requirements.txt
setuptools
datasets 
langchain 
PyPDF2
python-dotenv
pinecone-client
langchain-pinecone
transformers
-e .